[unit-test] enable tiff test
