+++ test/unit/imaging/tiff_io.cpp
@@ -1,7 +1,7 @@
 // disabled on windows due to https://github.com/mapnik/mapnik/issues/2838
 // TODO - get to the bottom of why including `tiff_reader.cpp` breaks windows
 // or re-write image_readers to allow `#include tiff_reader.hpp`
-#if !defined(_MSC_VER) && defined(HAVE_TIFF)
+#ifdef HAVE_TIFF
 
 #include "catch.hpp"
 
@@ -12,7 +12,7 @@
 #include <mapnik/util/file_io.hpp>
 #include <mapnik/util/fs.hpp>
 #include <boost/interprocess/streams/bufferstream.hpp>
-#include "../../../src/tiff_reader.cpp"
+#include "../../../src/tiff_reader.hpp"
 
 #if defined(MAPNIK_MEMORY_MAPPED_FILE)
 using source_type = boost::interprocess::ibufferstream;
